swap: reorganize swap tests locations
