+++ swap/swap.go
@@ -20,7 +20,6 @@ import (
 	"context"
 	"crypto/ecdsa"
 	"encoding/hex"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"math/big"
@@ -426,164 +425,6 @@ func (s *Swap) processAndVerifyCheque(cheque *Cheque, p *Peer) (uint64, error) {
 	}
 
 	return actualAmount, nil
-}
-
-// Balance returns the balance for a given peer
-func (s *Swap) Balance(peer enode.ID) (balance int64, err error) {
-	if swapPeer := s.getPeer(peer); swapPeer != nil {
-		return swapPeer.getBalance(), nil
-	}
-	err = s.store.Get(balanceKey(peer), &balance)
-	return balance, err
-}
-
-// Balances returns the balances for all known SWAP peers
-func (s *Swap) Balances() (map[enode.ID]int64, error) {
-	balances := make(map[enode.ID]int64)
-
-	s.peersLock.Lock()
-	for peer, swapPeer := range s.peers {
-		swapPeer.lock.Lock()
-		balances[peer] = swapPeer.getBalance()
-		swapPeer.lock.Unlock()
-	}
-	s.peersLock.Unlock()
-
-	// add store balances, if peer was not already added
-	balanceIterFunction := func(key []byte, value []byte) (stop bool, err error) {
-		peer := keyToID(string(key), balancePrefix)
-		if _, peerHasBalance := balances[peer]; !peerHasBalance {
-			var peerBalance int64
-			err = json.Unmarshal(value, &peerBalance)
-			if err == nil {
-				balances[peer] = peerBalance
-			}
-		}
-		return stop, err
-	}
-	err := s.store.Iterate(balancePrefix, balanceIterFunction)
-	if err != nil {
-		return nil, err
-	}
-
-	return balances, nil
-}
-
-// Cheques returns all known last sent and received cheques, grouped by peer
-func (s *Swap) Cheques() (map[enode.ID]*PeerCheques, error) {
-	cheques := make(map[enode.ID]*PeerCheques)
-
-	// get peer cheques from memory
-	s.peersLock.Lock()
-	for peer, swapPeer := range s.peers {
-		swapPeer.lock.Lock()
-		sentCheque := swapPeer.getLastSentCheque()
-		receivedCheque := swapPeer.getLastReceivedCheque()
-		// don't add peer to result if there are no cheques
-		if sentCheque != nil || receivedCheque != nil {
-			cheques[peer] = &PeerCheques{sentCheque, receivedCheque}
-		}
-		swapPeer.lock.Unlock()
-	}
-	s.peersLock.Unlock()
-
-	// get peer cheques from store
-	err := s.addStoreCheques(sentChequePrefix, cheques)
-	if err != nil {
-		return nil, err
-	}
-	err = s.addStoreCheques(receivedChequePrefix, cheques)
-	if err != nil {
-		return nil, err
-	}
-
-	return cheques, nil
-}
-
-// AvailableBalance returns the total balance of the chequebook against which new cheques can be written
-func (s *Swap) AvailableBalance() (uint64, error) {
-	// get the LiquidBalance of the chequebook
-	liquidBalance, err := s.contract.LiquidBalance(nil)
-	if err != nil {
-		return 0, err
-	}
-
-	// get all cheques
-	cheques, err := s.Cheques()
-	if err != nil {
-		return 0, err
-	}
-
-	// Compute the total worth of cheques sent and how much of of this is cashed
-	var sentChequesWorth uint64
-	var cashedChequesWorth uint64
-	for _, peerCheques := range cheques {
-		sentCheque := peerCheques.LastSentCheque
-		if sentCheque == nil {
-			continue
-		}
-		sentChequesWorth += sentCheque.ChequeParams.CumulativePayout
-		paidOut, err := s.contract.PaidOut(nil, sentCheque.ChequeParams.Beneficiary)
-		if err != nil {
-			return 0, err
-		}
-		cashedChequesWorth += paidOut.Uint64()
-	}
-	return liquidBalance.Uint64() + cashedChequesWorth - sentChequesWorth, nil
-}
-
-// add cheques from store for peers not already present in given cheques map
-func (s *Swap) addStoreCheques(chequePrefix string, cheques map[enode.ID]*PeerCheques) error {
-	chequesIterFunction := func(key []byte, value []byte) (stop bool, err error) {
-		peer := keyToID(string(key), chequePrefix)
-		// create struct if peer has no cheques entry yet
-		if peerCheques := cheques[peer]; peerCheques == nil {
-			cheques[peer] = &PeerCheques{}
-		}
-
-		// add cheque from store if not already in result
-		var peerCheque Cheque
-		err = json.Unmarshal(value, &peerCheque)
-		if err == nil {
-			switch chequePrefix {
-			case sentChequePrefix:
-				cheques[peer].LastSentCheque = &peerCheque
-			case receivedChequePrefix:
-				cheques[peer].LastReceivedCheque = &peerCheque
-			default:
-				err = fmt.Errorf("unknown type of cheque requested through prefix %s", chequePrefix)
-			}
-		}
-		return stop, err
-	}
-	return s.store.Iterate(chequePrefix, chequesIterFunction)
-}
-
-// PeerCheques contains the last cheque known to have been sent to a peer, as well as the last one received from the peer
-type PeerCheques struct {
-	LastSentCheque     *Cheque
-	LastReceivedCheque *Cheque
-}
-
-// PeerCheques returns the last sent and received cheques for a given peer
-func (s *Swap) PeerCheques(peer enode.ID) (PeerCheques, error) {
-	var sentCheque, receivedCheque *Cheque
-
-	swapPeer := s.getPeer(peer)
-	if swapPeer != nil {
-		sentCheque = swapPeer.getLastSentCheque()
-		receivedCheque = swapPeer.getLastReceivedCheque()
-	} else {
-		errSentCheque := s.store.Get(sentChequeKey(peer), &sentCheque)
-		if errSentCheque != nil && errSentCheque != state.ErrNotFound {
-			return PeerCheques{}, errSentCheque
-		}
-		errReceivedCheque := s.store.Get(receivedChequeKey(peer), &receivedCheque)
-		if errReceivedCheque != nil && errReceivedCheque != state.ErrNotFound {
-			return PeerCheques{}, errReceivedCheque
-		}
-	}
-	return PeerCheques{sentCheque, receivedCheque}, nil
 }
 
 // loadLastReceivedCheque loads the last received cheque for the peer from the store
